Renamed package to follow naming conventions
