+++ src/main/java/org/example/locators_page_object/homePage.java
@@ -1,4 +1,4 @@
-package org.example.locatorsPageObject;
+package org.example.locators_page_object;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
